+++ src/pyedb/dotnet/edb_core/cell/hierarchy/pin_pair_model.py
