+++ src/pyedb/dotnet/edb_core/cell/hierarchy/spice_model.py
