+++ src/pyedb/dotnet/edb_core/cell/terminal/bundle_terminal.py
@@ -20,8 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from pyedb.dotnet.edb_core.terminal.edge_terminal import EdgeTerminal
-from pyedb.dotnet.edb_core.terminal.terminal import Terminal
+from pyedb.dotnet.edb_core.cell.terminal.edge_terminal import EdgeTerminal
+from pyedb.dotnet.edb_core.cell.terminal.terminal import Terminal
 
 
 class BundleTerminal(Terminal):
