+++ src/pyedb/dotnet/edb_core/cell/terminal/terminal.py
